CSV data is UTC, Video is Localtime.  Make everything localtime.
My logic is that since this is a consumer app that only ever lives on one machine there is no
reason to ever, even accidentally, show the user a timestamp in anything but their timezone.
So we use the user's TZ everywhere.  If they move timezones, or if the race track is
in another timezone from their PC this may break.  In that case we need to do
a lot more diligence about collecting TZs and not just assuming an unlabeled timestamp
is localtz, which is hard and just out of scope / making my head hurt.